Move PLONK integration tests into the tests subdirectory.
